pyproject.toml
[tool.poetry]
name = "avatar-initials-py"
version = "0.1.0"
description = "Generate simple and unique avatars using a person's initials."
license = "MIT"
authors = ["Jonathan Hiles <jonathan@hil.es>"]
readme = "README.md"
repository = "https://github.com/axieum/avatar-initials-py"
keywords = ["avatar", "image", "initials", "person", "pillow", "text"]
classifiers = [
  "Intended Audience :: Developers",
  "Intended Audience :: End Users/Desktop",
  "Operating System :: OS Independent",
  "Topic :: Multimedia :: Graphics",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Typing :: Typed",
]
packages = [{ include = "avatar_initials_py", from = "src" }]
include = [{ path = "tests", format = "sdist" }]

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^9.4.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
pytest-xdist = "^3.1.0"
syrupy = "^3.0.5"

[tool.poetry.group.typing.dependencies]
mypy = ">=0.990"
types-pillow = "^9.3.0.4"

[tool.black]
target-version = ["py37"]
line_length = 120

[tool.isort]
py_version = 37
profile = "black"
src_paths = ["src", "tests"]
combine_as_imports = true

[tool.mypy]
files = ["src", "tests"]
strict = true

[tool.pytest.ini_options]
addopts = "-n auto"
testpaths = ["tests"]

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "if TYPE_CHECKING:",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
